Rename package to mfatypes.
