+++ lib/auth/sso_mfa_test.go
@@ -34,7 +34,7 @@ import (
 	mfav1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/mfa/v1"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/auth/authclient"
-	"github.com/gravitational/teleport/lib/auth/mfa"
+	"github.com/gravitational/teleport/lib/auth/mfatypes"
 	"github.com/gravitational/teleport/lib/authz"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/services"
@@ -276,7 +276,7 @@ func TestSSOMFAChallenge_Creation(t *testing.T) {
 					Username:      samlUser.GetName(),
 					ConnectorID:   samlConnector.GetName(),
 					ConnectorType: samlConnector.GetKind(),
-					ChallengeExtensions: &mfa.ChallengeExtensions{
+					ChallengeExtensions: &mfatypes.ChallengeExtensions{
 						Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 					},
 				}, sd)
@@ -315,7 +315,7 @@ func TestSSOMFAChallenge_Creation(t *testing.T) {
 					Username:      oidcUser.GetName(),
 					ConnectorID:   oidcConnector.GetName(),
 					ConnectorType: oidcConnector.GetKind(),
-					ChallengeExtensions: &mfa.ChallengeExtensions{
+					ChallengeExtensions: &mfatypes.ChallengeExtensions{
 						Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 					},
 				}, sd)
@@ -481,7 +481,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      "wrong-user",
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 				},
 				Token: "token",
@@ -505,7 +505,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      samlUser.GetName(),
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 				},
 				Token: "token",
@@ -529,7 +529,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      samlUser.GetName(),
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 				},
 			},
@@ -552,7 +552,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      samlUser.GetName(),
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 				},
 				Token: "token",
@@ -576,7 +576,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      samlUser.GetName(),
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope:      mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 					AllowReuse: mfav1.ChallengeAllowReuse_CHALLENGE_ALLOW_REUSE_YES,
 				},
@@ -602,7 +602,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      noMFASAMLUser.GetName(),
 				ConnectorID:   noMFASAMLConnector.GetName(),
 				ConnectorType: noMFASAMLConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 				},
 				Token: "token",
@@ -626,7 +626,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      standardUser.GetName(),
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope: mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 				},
 				Token: "token",
@@ -650,7 +650,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      samlUser.GetName(),
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope:      mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 					AllowReuse: mfav1.ChallengeAllowReuse_CHALLENGE_ALLOW_REUSE_NO,
 				},
@@ -680,7 +680,7 @@ func TestSSOMFAChallenge_Validation(t *testing.T) {
 				Username:      samlUser.GetName(),
 				ConnectorID:   samlConnector.GetName(),
 				ConnectorType: samlConnector.GetKind(),
-				ChallengeExtensions: &mfa.ChallengeExtensions{
+				ChallengeExtensions: &mfatypes.ChallengeExtensions{
 					Scope:      mfav1.ChallengeScope_CHALLENGE_SCOPE_LOGIN,
 					AllowReuse: mfav1.ChallengeAllowReuse_CHALLENGE_ALLOW_REUSE_YES,
 				},
