+++ lib/auth/webauthn/login.go
@@ -35,7 +35,7 @@ import (
 
 	mfav1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/mfa/v1"
 	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/lib/auth/mfa"
+	"github.com/gravitational/teleport/lib/auth/mfatypes"
 	wantypes "github.com/gravitational/teleport/lib/auth/webauthntypes"
 )
 
@@ -191,7 +191,7 @@ func (f *loginFlow) begin(ctx context.Context, user string, challengeExtensions 
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
-	sd.ChallengeExtensions = &mfa.ChallengeExtensions{
+	sd.ChallengeExtensions = &mfatypes.ChallengeExtensions{
 		Scope:                       challengeExtensions.Scope,
 		AllowReuse:                  challengeExtensions.AllowReuse,
 		UserVerificationRequirement: challengeExtensions.UserVerificationRequirement,
