+++ lib/services/sso_mfa.go
@@ -18,7 +18,7 @@
 
 package services
 
-import "github.com/gravitational/teleport/lib/auth/mfa"
+import "github.com/gravitational/teleport/lib/auth/mfatypes"
 
 // SSOMFASessionData SSO MFA Session data.
 type SSOMFASessionData struct {
@@ -34,5 +34,5 @@ type SSOMFASessionData struct {
 	// ConnectorType is SSO type of the corresponding Auth connector (SAML, OIDC).
 	ConnectorType string `json:"connector_type,omitempty"`
 	// ChallengeExtensions are Teleport extensions that apply to this SSO MFA session.
-	ChallengeExtensions *mfa.ChallengeExtensions `json:"challenge_extensions"`
+	ChallengeExtensions *mfatypes.ChallengeExtensions `json:"challenge_extensions"`
 }
